Remove next/app from appWithI18n + refactor
