+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "next-translate",
-  "version": "1.0.0-experimental.14",
+  "version": "1.0.0-experimental.15",
   "description": "Next.js utility to translate pages in a easy way.",
   "license": "MIT",
   "keywords": [
